Removed, renamed and refactored
