--- frontend/src/components/regions/AgileBoard/AgileCard/AgileCardActions/index.js
@@ -1,61 +1,0 @@
-import React from "react";
-import Presentation from "./Presentation";
-import { showButtons } from "../../../../../utils/cardButtons";
-
-export default ({
-  authUser,
-  viewedUser,
-
-  card,
-  handleClickAddReview,
-  handleClickOpenCardDetails,
-
-  handleRequestReview,
-  handleStartProject,
-  handleCancelReviewRequest,
-
-  handleClickOpenWorkshopAttendanceForm,
-  handleStartTopic,
-  handleStopTopic,
-  handleFinishTopic,
-  handleRemoveWorkshopAttendance,
-  loadingStartProject,
-  loadingStartTopic,
-  loadingClickOpenWorkshopAttendanceForm,
-  loadingRequestReview,
-  loadingCancelReviewRequest,
-  loadingStopTopic,
-  loadingFinishTopic,
-  loadingRemoveWorkshopAttendance,
-}) => {
-  const props = {
-    card,
-    ...showButtons({
-      card,
-      authUser,
-      viewedUser,
-    }),
-    handleClickAddReview,
-    handleClickOpenCardDetails,
-
-    handleRequestReview,
-    handleStartProject,
-    handleCancelReviewRequest,
-
-    handleClickOpenWorkshopAttendanceForm,
-    handleStartTopic,
-    handleStopTopic,
-    handleFinishTopic,
-    handleRemoveWorkshopAttendance,
-
-    loadingStartProject,
-    loadingStartTopic,
-    loadingClickOpenWorkshopAttendanceForm,
-    loadingRequestReview,
-    loadingCancelReviewRequest,
-    loadingStopTopic,
-    loadingFinishTopic,
-    loadingRemoveWorkshopAttendance,
-  };
-  return <Presentation {...props} />;
-};
