--- frontend/src/components/regions/AgileBoard/AgileCard/utils.js
@@ -1,7 +1,0 @@
-export function checkIfCardIsInReviewColumn({ card }) {
-  return card.status === "IR";
-}
-
-export function userReviewedSinceLastReviewRequest({ viewedUser, card }) {
-  return card.usersThatReviewedSinceLastReviewRequest.includes(viewedUser.id);
-}
