--- frontend/src/components/regions/AgileBoard/mockedData.js
@@ -1,257 +1,0 @@
-export const cards = {
-  3: {
-    id: 3,
-    contentItem: 19,
-    contentItemUrl: "http://syllabus.africacode.net/topics/clean-code/python/",
-    status: "C",
-    recruitProject: 2,
-    assignees: [4, 5, 2],
-    reviewers: [4, 5, 2],
-    assigneeNames: [
-      "vmeteni@gmail.com",
-      "vr.meteni@gmail.com",
-      "vuyisanani.meteni@umuzi.org",
-    ],
-    reviewerNames: [
-      "vmeteni@gmail.com",
-      "vr.meteni@gmail.com",
-      "vuyisanani.meteni@umuzi.org",
-    ],
-    isHardMilestone: false,
-    isSoftMilestone: false,
-    title: "Clean Code for Python",
-    contentType: "T",
-    contentTypeNice: "topic",
-    tagNames: ["python", "clean-code"],
-    order: 0,
-    codeReviewCompetentSinceLastReviewRequest: 0,
-    codeReviewExcellentSinceLastReviewRequest: 0,
-    codeReviewRedFlagSinceLastReviewRequest: 0,
-    codeReviewNyCompetentSinceLastReviewRequest: 0,
-    requiresCards: [1],
-    requiredByCards: [4, 5],
-    flavourNames: ["a", "b"],
-    projectSubmissionTypeNice: null,
-    topicNeedsReview: false,
-    topicProgress: 1,
-    dueTime: "2021-10-13T21:03:10Z",
-    completeTime: "2021-10-13T21:03:05Z",
-    reviewRequestTime: "2021-10-13T21:03:06Z",
-    startTime: "2021-10-13T21:03:08Z",
-    canStart: false,
-    canForceStart: false,
-    openPrCount: 0,
-    oldestOpenPrUpdatedTime: null,
-    repoUrl: "Vuyisanan",
-    usersThatReviewedSinceLastReviewRequest: [],
-  },
-  4: {
-    id: 4,
-    contentItem: 33,
-    contentItemUrl:
-      "http://syllabus.africacode.net/language-agnostic/validate-id-number/",
-    status: "R",
-    recruitProject: 5,
-    assignees: [2],
-    reviewers: [3, 4, 5, 13, 2],
-    assigneeNames: ["vuyisanani.meteni@umuzi.org"],
-    reviewerNames: [
-      "someone.nice@example.com",
-      "vmeteni@gmail.com",
-      "vr.meteni@gmail.com",
-      "foo.8@example.com",
-      "vuyisanani.meteni@umuzi.org",
-    ],
-    isHardMilestone: false,
-    isSoftMilestone: false,
-    title: "Validate a South African ID number",
-    contentType: "P",
-    contentTypeNice: "project",
-    tagNames: ["tdd", "problem-solving", "validation"],
-    order: 0,
-    codeReviewCompetentSinceLastReviewRequest: 0,
-    codeReviewExcellentSinceLastReviewRequest: 0,
-    codeReviewRedFlagSinceLastReviewRequest: 0,
-    codeReviewNyCompetentSinceLastReviewRequest: 0,
-    requiresCards: [1, 2, 3],
-    requiredByCards: [5, 6],
-    flavourNames: ["test"],
-    projectSubmissionTypeNice: "repo",
-    topicNeedsReview: false,
-    topicProgress: 2,
-    dueTime: "2021-10-14T07:48:08Z",
-    completeTime: "2021-10-14T07:48:02Z",
-    reviewRequestTime: "2021-10-14T07:48:04Z",
-    startTime: "2021-10-14T07:48:05Z",
-    canStart: true,
-    canForceStart: true,
-    openPrCount: 0,
-    oldestOpenPrUpdatedTime: null,
-    repoUrl: null,
-    usersThatReviewedSinceLastReviewRequest: [],
-  },
-  6: {
-    id: 6,
-    contentItem: 23,
-    contentItemUrl: "http://syllabus.africacode.net/projects/katas/level-2/",
-    status: "C",
-    recruitProject: 6,
-    assignees: [2],
-    reviewers: [4],
-    assigneeNames: ["vuyisanani.meteni@umuzi.org"],
-    reviewerNames: ["vmeteni@gmail.com"],
-    isHardMilestone: false,
-    isSoftMilestone: false,
-    title: "Level 2 coding challenges",
-    contentType: "P",
-    contentTypeNice: "project",
-    tagNames: ["clean-code", "algrithms"],
-    order: 0,
-    codeReviewCompetentSinceLastReviewRequest: 1,
-    codeReviewExcellentSinceLastReviewRequest: 0,
-    codeReviewRedFlagSinceLastReviewRequest: 0,
-    codeReviewNyCompetentSinceLastReviewRequest: 0,
-    requiresCards: [4],
-    requiredByCards: [],
-    flavourNames: ["userTest"],
-    projectSubmissionTypeNice: "continue_repo",
-    topicNeedsReview: false,
-    topicProgress: 5,
-    dueTime: null,
-    completeTime: "2021-10-21T20:53:30.772712Z",
-    reviewRequestTime: null,
-    startTime: null,
-    canStart: false,
-    canForceStart: false,
-    openPrCount: 0,
-    oldestOpenPrUpdatedTime: null,
-    repoUrl: "Vuyisanan",
-    usersThatReviewedSinceLastReviewRequest: [],
-  },
-  7: {
-    id: 7,
-    contentItem: 4,
-    contentItemUrl:
-      "http://syllabus.africacode.net/topics/bootcamp/failure-at-bootcamp/",
-    status: "B",
-    recruitProject: null,
-    assignees: [2],
-    reviewers: [4],
-    assigneeNames: ["vuyisanani.meteni@umuzi.org"],
-    reviewerNames: ["vmeteni@gmail.com"],
-    isHardMilestone: false,
-    isSoftMilestone: false,
-    title: "SUPER IMPORTANT STUFF: how to not fail",
-    contentType: "T",
-    contentTypeNice: "topic",
-    tagNames: [],
-    order: 0,
-    codeReviewCompetentSinceLastReviewRequest: 0,
-    codeReviewExcellentSinceLastReviewRequest: 0,
-    codeReviewRedFlagSinceLastReviewRequest: 0,
-    codeReviewNyCompetentSinceLastReviewRequest: 0,
-    requiresCards: [5],
-    requiredByCards: [],
-    flavourNames: ["testIn"],
-    projectSubmissionTypeNice: null,
-    topicNeedsReview: false,
-    topicProgress: 6,
-    dueTime: "2021-10-20T06:38:14Z",
-    completeTime: null,
-    reviewRequestTime: null,
-    startTime: "2021-10-20T06:38:17Z",
-    canStart: false,
-    canForceStart: true,
-    repoUrl: null,
-    usersThatReviewedSinceLastReviewRequest: [],
-  },
-  8: {
-    id: 8,
-    contentItem: 24,
-    contentItemUrl:
-      "http://syllabus.africacode.net/language-agnostic/problem-solving-technique/",
-    status: "IP",
-    recruitProject: 7,
-    assignees: [5],
-    reviewers: [2],
-    assigneeNames: ["vr.meteni@gmail.com"],
-    reviewerNames: ["vuyisanani.meteni@umuzi.org"],
-    isHardMilestone: true,
-    isSoftMilestone: true,
-    title: "Problem decomposition",
-    contentType: "T",
-    contentTypeNice: "topic",
-    tagNames: ["problem solving"],
-    order: 0,
-    codeReviewCompetentSinceLastReviewRequest: 0,
-    codeReviewExcellentSinceLastReviewRequest: 0,
-    codeReviewRedFlagSinceLastReviewRequest: 0,
-    codeReviewNyCompetentSinceLastReviewRequest: 0,
-    requiresCards: [5],
-    requiredByCards: [],
-    flavourNames: ["new"],
-    projectSubmissionTypeNice: null,
-    topicNeedsReview: false,
-    topicProgress: 7,
-    dueTime: null,
-    completeTime: null,
-    reviewRequestTime: null,
-    startTime: null,
-    canStart: false,
-    canForceStart: false,
-    openPrCount: 1,
-    oldestOpenPrUpdatedTime: "2021-10-14T07:35:41.774375Z",
-    repoUrl: "https://something.com/like_this4",
-    usersThatReviewedSinceLastReviewRequest: [],
-  },
-};
-export const latestCalls = {
-  B: {
-    anyLoading: false,
-    lastAssigneeCallPage: 2,
-    lastReviewerCallPage: 0,
-    assigneeCallResponseCount: 0,
-    reviewerCallResponseCount: -1,
-    totalCards: 1,
-  },
-  R: {
-    anyLoading: false,
-    lastAssigneeCallPage: 2,
-    lastReviewerCallPage: 0,
-    assigneeCallResponseCount: 0,
-    reviewerCallResponseCount: -1,
-    totalCards: 1,
-  },
-  IP: {
-    anyLoading: false,
-    lastAssigneeCallPage: 1,
-    lastReviewerCallPage: 1,
-    assigneeCallResponseCount: 0,
-    reviewerCallResponseCount: 1,
-    totalCards: 1,
-  },
-  RF: {
-    anyLoading: false,
-    lastAssigneeCallPage: 1,
-    lastReviewerCallPage: 1,
-    assigneeCallResponseCount: 0,
-    reviewerCallResponseCount: 0,
-    totalCards: 0,
-  },
-  IR: {
-    anyLoading: false,
-    lastAssigneeCallPage: 1,
-    lastReviewerCallPage: 1,
-    assigneeCallResponseCount: 0,
-    reviewerCallResponseCount: 0,
-    totalCards: 0,
-  },
-  C: {
-    anyLoading: false,
-    lastAssigneeCallPage: 2,
-    lastReviewerCallPage: 0,
-    assigneeCallResponseCount: 0,
-    reviewerCallResponseCount: -1,
-    totalCards: 2,
-  },
-};
