--- frontend/src/components/regions/AgileBoard/redux/actions.js
@@ -1,3 +1,0 @@
-// import types from "./types.js";
-
-export default {};
