--- frontend/src/components/regions/AgileBoard/redux/index.js
@@ -1,5 +1,0 @@
-import agileBoardReducer from "./reducers";
-export { default as agileBoardOperations } from "./operations";
-export { default as agileBoardTypes } from "./types";
-export { default as agileBoardSagas } from "./sagas";
-export default agileBoardReducer;
