--- frontend/src/components/regions/AgileBoard/redux/operations.js
@@ -1,5 +1,0 @@
-import Creators from "./actions.js";
-
-export default {
-  ...Creators,
-};
