--- frontend/src/components/regions/AgileBoard/redux/reducers.js
@@ -1,9 +1,0 @@
-// import types from "./types";
-const INITIAL_STATE = {};
-const reducer = (state = INITIAL_STATE, action) => {
-  switch (action.type) {
-    default:
-      return state;
-  }
-};
-export default reducer;
