--- frontend/src/components/regions/AgileBoard/redux/sagas.js
@@ -1,6 +1,0 @@
-// import { takeEvery, put } from "redux-saga/effects";
-// import { apiReduxApps } from "../../../../apiAccess/apiApps";
-
-// import constants from "../../../../constants";
-
-export default [];
