--- frontend/src/components/regions/AgileBoard/redux/types.js
@@ -1,1 +1,0 @@
-export default {};
