+++ frontend/src/stories/AgileCardActions.stories.js
@@ -1,9 +1,9 @@
 import React from "react";
-import AgileCardActions from "../components/regions/AgileBoard/AgileCard/AgileCardActions/Presentation";
+import AgileCardActions from "../components/pages/AgileBoard/AgileCard/AgileCardActions/Presentation";
 import card from "../stories/fixtures/AgileCards/openPrCard.json";
 
 export default {
-    title: "Tilde/AgileCardActions",
+    title: "Tilde/pages/AgileCardActions",
     component: AgileCardActions,
 }
